fix: Remove string pop + pop before block entry
Removed the pop on the :string state that is only used as mixin and thus
causes invalid state transitions (we traverse back too far) -- pop! may
have been useful if we pushed into the :string state but since we mixin
the :string state there is no need.

Pop out of :construct_arg upon detection of a "{" and before traversing
into the :construct_body state. This may be redesigned again as this the
pop before push seems a bit smelly, although the double pop upon "}" was
also a questionable design choice. WIP